Xcode warnings fix

+++ src/game/g_client.c
@@ -1160,9 +1160,9 @@ char *ClientConnect( int clientNum, qboolean firstTime, qboolean isBot ) {
 	}
 
 	// Gordon: porting q3f flag bug fix
-	//			If a player reconnects quickly after a disconnect, the client disconnect may never be called, thus flag can get lost in the ether
+	// If a player reconnects quickly after a disconnect, the client disconnect may never be called, thus flag can get lost in the ether
 	if ( ent->inuse ) {
-		G_LogPrintf( "Forcing disconnect on active client: %i\n", ent - g_entities );
+		G_LogPrintf( "Forcing disconnect on active client: %d\n", (int)(ent - g_entities) );
 		// so lets just fix up anything that should happen on a disconnect
 		ClientDisconnect( ent - g_entities );
 	}
