+++ src/game/g_trigger.c
@@ -1442,7 +1442,7 @@ void SP_trigger_objective_info( gentity_t *ent ) {
 		}
 	}
 
-	G_SetConfigStringValue( CS_OID_DATA + level.numOidTriggers, "e", va( "%i", ent - g_entities )           );
+	G_SetConfigStringValue( CS_OID_DATA + level.numOidTriggers, "e", va( "%d", (int)(ent - g_entities) )           );
 	G_SetConfigStringValue( CS_OID_DATA + level.numOidTriggers, "o", va( "%i", objflags )             );
 	G_SetConfigStringValue( CS_OID_DATA + level.numOidTriggers, "cix",   va( "%i", cix )                      );
 	G_SetConfigStringValue( CS_OID_DATA + level.numOidTriggers, "cia",   va( "%i", cia )                      );
@@ -1501,7 +1501,7 @@ void trigger_concussive_touch( gentity_t *ent, gentity_t *other, trace_t *trace 
 	trace = trace;
 
 	return; // FIXME this should be NULLed out in SP_trigger_concussive_dust after everything works
-	G_Printf( "hit concussive ent %d mins=%f,%f,%f maxs=%f,%f,%f\n",ent - g_entities,
+	G_Printf( "hit concussive ent %d mins=%f,%f,%f maxs=%f,%f,%f\n", (int)(ent - g_entities),
 			  ent->r.mins[0],
 			  ent->r.mins[1],
 			  ent->r.mins[2],
